requirements.txt
asgiref==3.5.2
certifi==2022.12.7
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1.1
django-ckeditor==6.5.0
django-excel==0.0.10
django-filter==22.1
django-guardian==2.4.0
django-heroku==0.3.1
django-js-asset==2.0.0
django-model-utils==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
Flask==2.2.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lml==0.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.4
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-ods==0.6.0
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
pyexcel-xlsxw==0.6.1
PyJWT==2.6.0
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.2
texttable==1.6.4
twilio==7.15.4
tzdata==2022.2
urllib3==1.26.13
Werkzeug==2.2.2
whitenoise==6.2.0
XlsxWriter==3.0.3
xmltodict==0.13.0
appdirs==1.4.3

attrs==16.3.0
autobahn==0.17.2
Automat==0.5.0
channels==1.0.3
constantly==15.1.0
daphne==1.0.3

incremental==16.10.1
oauthlib==2.0.1
packaging==16.8
pyparsing==2.2.0

requests-oauthlib==0.8.0
six==1.10.0
Twisted==17.1.0
txaio==2.6.1
typing==3.5.3.0
zope.interface==4.3.3